updated frontend

+++ frontend/src/components/TaskList.js
@@ -13,26 +13,34 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const react_1 = require("@chakra-ui/react");
-const useContactsContext_1 = require("../hooks/useContactsContext");
+const useTasksContext_1 = require("../hooks/useTasksContext");
 const react_2 = __importDefault(require("react"));
-const ContactList = ({ contact }) => {
-    const { dispatch } = (0, useContactsContext_1.useContactsContext)();
+const useAuthContext_1 = require("../hooks/useAuthContext");
+const TaskList = ({ task }) => {
+    const { dispatch } = (0, useTasksContext_1.useTasksContext)();
+    const { user } = (0, useAuthContext_1.useAuthContext)();
     const handleClick = () => __awaiter(void 0, void 0, void 0, function* () {
-        const response = yield fetch('/api/contacts/' + contact._id, {
-            method: 'DELETE'
+        if (!user) {
+            return;
+        }
+        const response = yield fetch('/api/tasks/' + task._id, {
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
         const json = yield response.json();
         if (response.ok) {
-            dispatch({ type: 'DELETE_CONTACT', payload: json });
+            dispatch({ type: 'DELETE_TASK', payload: json });
         }
     });
     return (react_2.default.createElement(react_1.SimpleGrid, { p: "20px" },
-        react_2.default.createElement(react_1.Card, { align: "left", width: "500px", borderRadius: 20, boxShadow: "lg", bg: "gray.50", p: "20px" },
+        react_2.default.createElement(react_1.Card, { align: "left", width: "50%", borderRadius: 20, boxShadow: "lg", bg: "gray.50", p: "20px" },
             react_2.default.createElement(react_1.Flex, { alignItems: "center", width: "100%" },
-                react_2.default.createElement(react_1.Heading, { color: "blue.500" }, contact.name),
+                react_2.default.createElement(react_1.Heading, { color: "blue.500" }, task.name),
                 react_2.default.createElement(react_1.Spacer, null),
                 react_2.default.createElement(react_1.Button, { onClick: handleClick }, "Delete")),
-            react_2.default.createElement(react_1.Text, { textAlign: "left" }, contact.number),
-            react_2.default.createElement(react_1.Text, { textAlign: "left" }, contact.email))));
+            react_2.default.createElement(react_1.Text, { textAlign: "left" }, task.subject),
+            react_2.default.createElement(react_1.Text, { textAlign: "left" }, task.description))));
 };
-exports.default = ContactList;
+exports.default = TaskList;
